+++ examples/java-http-server/src/main/java/com/twitter/hello/server/HelloWorldException.java
@@ -1,0 +1,8 @@
+package com.twitter.hello.server;
+
+public class HelloWorldException extends RuntimeException {
+
+    public HelloWorldException(String message) {
+        super(message);
+    }
+}