+++ http/src/main/scala/com/twitter/finatra/http/internal/exceptions/AbstractFrameworkExceptionMapper.scala
@@ -4,7 +4,7 @@ import com.twitter.finagle.http.{Request, Response}
 import com.twitter.finatra.http.exceptions.ExceptionMapper
 import com.twitter.finatra.http.response.ResponseBuilder
 
-private[exceptions] abstract class AbstractExceptionMapper[T <: Throwable](
+private[exceptions] abstract class AbstractFrameworkExceptionMapper[T <: Throwable](
   response: ResponseBuilder)
   extends ExceptionMapper[T] {