+++ http/src/test/java/com/twitter/finatra/http/tests/integration/doeverything/main/DoEverythingJavaServerMain.java
@@ -1,4 +1,4 @@
-package com.twitter.finatra.http.tests.integration.main;
+package com.twitter.finatra.http.tests.integration.doeverything.main;
 
 public final class DoEverythingJavaServerMain {
     private DoEverythingJavaServerMain() {